cleanup: remove getOssVolumeOptions for CreateVolume to ensure the mount options for dynamic and static volumes are same
